[distributor]: return instance limit errors with HTTP status code 503 (#13032)
#### What this PR does
This PR updates the HTTP status code used for distributors’ instance
limit errors from 500 Internal Server Error to 503 Service Unavailable.
This change aligns the HTTP status code with the existing gRPC code
Unavailable, as introduced in
https://github.com/grafana/mimir/pull/13003.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Yuri Nikolic <durica.nikolic@grafana.com>