index planning: fix planner creation
we were skipping creating planners for on-disk blocks. This happened because planners are created for each block upon opening the prometheus TSDB. The callback to create a planner needed to have the userDB created and saved in the Ingester. The userDB wasn't saved because it was still being created.

This PR moves the planer creation to the userDB instead of the ingester.

this was added as a bug in XXX

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>