Admin UI: use relative links (#13034)
#### What this PR does

Changes the links and static sources to use relative links, instead of
absolute ones in the UI.

This helps when the UI is served through a reverse-proxy.

#### Which issue(s) this PR fixes or relates to

Fixes an internal issue.

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>