[distributor]: refactor reactive limiter permit checks (#13033)
#### What this PR does

`Distributor.startPushRequest()` performs several checks when handling a
new push request. One of these checks verifies whether the reactive
limiter can acquire a permit. The function also creates a `requestState`
object and stores it in the context. The presence of this object
indicates that a push request has already been started and is currently
in-flight.

Since `Distributor.startPushRequest()` may be called multiple times for
the same request, it should first checks whether a `requestState` object
already exists in the context. If it does, the function should
immediately return to avoid reinitializing the same request.

In https://github.com/grafana/mimir/pull/12923, the permit check for the
reactive limiter was mistakenly placed before the check for an existing
requestState object.

This PR corrects that error by enforcing that the `requestState` check
runs first.

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Yuri Nikolic <durica.nikolic@grafana.com>