Revise health check: implement CheckReady instead
In keeping with the other components (e.g, ingester), updates the
MultitenantCompactor so it implements the `CheckReady` function.  This
function runs the three operations: write, stat, read an empty file in
the /data volume's file system.

Updates `mimir.go` so it calls this function as part of the readiness
check.