Step invariant expression handling (#12931)
#### What this PR does

This PR adds support for handling step invariant expressions in the MQE
query planning and evaluation.

This includes;

* support for serialization of query plans with step invariant
expressions
* support for optimizing the query evaluation to query a single step
value and re-use this to populate each step in the result vector

#### Which issue(s) this PR fixes or relates to

Fixes https://github.com/grafana/mimir-squad/issues/3017

#### Checklist

- [x] Tests updated.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.

---------

Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>