index planning: fix planner creation (#13035)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

we were skipping creating planners for on-disk blocks. This happened
because planners are created for each block upon opening the prometheus
TSDB. The callback to create a planner needed to have the userDB created
and saved in the Ingester. The userDB wasn't saved because it was still
being created.

This PR moves the planer creation to the userDB instead of the ingester.

this was introduced in #12607 


#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>