Jsonnet: move multi_zone_availability_zones config out of multi-zone-distributor.libsonnet (#13044)
#### What this PR does

`multi_zone_availability_zones` is used by all Mimir components and it's
not distributor specific. In this I'm moving it out of
`multi-zone-distributor.libsonnet` and adding an assertion to ensure no
more than 3 AZs are specified.

_This PR is expected to be a no-op._

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Marco Pracucci <marco@pracucci.com>