Add store-gateway support for -querier.prefer-availability-zone (#13045)
#### What this PR does

We currently have `-querier.prefer-availability-zone` that allows to
specify the preferred zone when querying ingesters. In this PR I'm
extending it to store-gateways too. The CLI flag was experimental and
hidden. In this PR I'm exposing it in the doc, but it's still
experimental.

In this PR I've also added 2 new metrics, tracking the data transfer
between queriers and ingesters / store-gateways respectively:

- `cortex_ingester_client_transferred_bytes_total{ingester_zone="..."}`
-
`cortex_storegateway_client_transferred_bytes_total{store_gateway_zone="..."}`

These two metrics should make it easily to monitor the cross-AZ data
transfer that happens during fallbacks.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>