compactor: allow skipping jobs with blocks' maxTime in the future (#13040)
#### What this PR does

When the compactor runs against blocks produced from the block-builder,
we observed an elevated and unbalanced workload. We believe this's due
to the fact that the block-builder produces blocks during the whole
period of an hour. This pattern is very different from how ingesters
produce blocks, at (almost) exactly the beginning of every second hour.

In this PR I'm updating how the compactor skips the "too recent" blocks.
The theory is that this will allow making its workload more smooth (will
need to verify that against a dev cell, first).

The feature is controlled by the experimental
`-compactor.first-level-compaction-skip-future-max-time` (off by
default).

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>
Co-authored-by: Andy Asp <90626759+andyasp@users.noreply.github.com>