ingest storage: cap number of bytes requested from kafka
### Problem
it's possible that some fetches throw off our estimations for the bytes per record. The `fetchWant` will attempt to fetch the number of records it was given, but if `UpdateBytesPerRecord` was called by the fetcher worker in the snippet below with a very high average for each record, then it's possible that the `fetchWant` goes over its `targetMaxBytes`

https://github.com/grafana/mimir/blob/1fdeb12b59ffe9682c3eb6980d9c79f777c2b651/pkg/storage/ingest/fetcher.go#L762

### What this PR does

This PR renames the previous `targetMaxBytes` to `maxBytesLimit` and enforces it in invocations to `MaxBytes()`. That way we also cap the memory usage of the fetcher worker.