mimir-distributed: don't route ingress to headless services (#13064)
#### What this PR does

As discussed in the #10994, the use of headless services in the ingress
stays in the way of users. There is no known reason to do that, so we
agreed to make the change in the upcoming major release.

#### Which issue(s) this PR fixes or relates to

Fixes #10994

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>