chore: move file removal to deferred func
This updates the process which removes the compactor health-check file
into a deferred function so it is always run before returning.

Also tweaked the logging a little in this function: all returned errors
are logged regardless, so we don't log in the `CheckReady` function and
make sure to wrap any `-compactor.data-dir` `stat` errors before returning
them.