docs: fix image that was reverted by arch split (#13079)
#### What this PR does

This fixes the read path diagram which marked the query-scheduler as
required in #12609. It was incorrectly overwritten in #12806.

#### Which issue(s) this PR fixes or relates to

Related #12609

Related #12806

#### Checklist

- [ ] Tests updated.
- [X] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>
(cherry picked from commit 12ef9c85e34aa5082b1e43600fb32725dfdba7b8)