Hide max_active_series_per_user from the docs, and remove the TODO fallback (#13085)
#### What this PR does

Removes the new `max_active_series_per_user` field from the doc, as that
feature is experimental so far. Also removes the TODO fallback to the
`max_global_series_per_user` when it's unset.

#### Which issue(s) this PR fixes or relates to

Fixes docs

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>