Remote execution: querier backwards compatibility support (#13017)
#### What this PR does

This PR changes the behaviour of query-frontends to only generate query
plans that the running set of queriers support.

This ensures that query-frontends don't start generating plans that
queriers don't support during rollouts until all queriers have upgraded.

Queriers now join a ring and report the maximum supported query plan
version as part of their instance description, and query-frontends then
pass this to the query planner.

Optimisations passes can then choose to skip optimisations that
introduce functionality that is not supported by the running set of
queriers.

Not included in this PR, but will be added separately: alerts to warn
when the cluster is in an inconsistent state (eg. query-frontends are
reporting a maximum supported version that does not match what queriers
support).

Don't be put off by the large number of files changed: most of these are
the regenerated Helm or Jsonnet test files, or files that needed to be
updated due to the changed signature of `AddIngester` from the dskit PR.

:warning: This PR depends on https://github.com/grafana/dskit/pull/757.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12808,
https://github.com/grafana/mimir/pull/12961

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.