Remote execution: fix issue where requests that are cancelled or time out are cached (#13098)
#### What this PR does

This PR fixes an issue where queries evaluated with remote execution
that are cancelled or time out can be incorrectly cached by the error
caching middleware. This causes subsequent identical requests to also
return the same error.

This happened because requests that are cancelled or time out could
return a gRPC-style error, and these were translated to the default
"execution" error type by `apierror.TypeForError` when called by
`engineQueryRequestRoundTripperHandler`'s `convertToAPIError` method.

#### Which issue(s) this PR fixes or relates to

#12551

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.