Remote execution: improve logging of errors when streaming results from querier (#13099)
#### What this PR does

This PR improves the logging from queriers when sending a message to
query-frontends fails.

It also fixes an issue where a non-client-generated error (eg.
networking issue) while sending a message to query-frontends results in
queriers logging an "EOF" error. Instead, we need to call `RecvMsg()` to
discover the error.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12551

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.