Remote execution: fix panic in querier if writing message to stream fails (#13121)
#### What this PR does

This PR fixes an issue introduced in
https://github.com/grafana/mimir/pull/13099.

If writing a message to the stream fails and returns `EOF`, calling
`RecvMsg` with a `nil` target causes the querier to panic.

#### Which issue(s) this PR fixes or relates to

#13099

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.