Fix issue where log messages emitted by `queryResponseWriter` aren't attached to the request's trace (#13124)
#### What this PR does

This PR fixes an issue where the log messages emitted by
`queryResponseWriter` aren't attached to the request's trace.

I've chosen not to add a changelog entry given this is not a user-facing
change.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.