Promote rate limit logger configurations (#13128)
#### What this PR does

In https://github.com/grafana/mimir/pull/5764, we introduced rate
limiting for Mimir’s logger, allowing it to control the number of log
messages emitted per second. This functionality was initially released
as experimental and controlled via the following CLI flags:
```
-log.rate-limit-enabled
-log.rate-limit-logs-per-second and
-log.rate-limit-logs-per-second-burst
```

This PR promotes these CLI flags from **experimental** to **stable**.

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Yuri Nikolic <durica.nikolic@grafana.com>