ingester: mirror querier: ignore fresh series
Series created between running the first query with index planning and running the second query without index planning may show up in the result and create a discrepancy. I call these fresh series.

Detecting fresh series isn't trivial because of out-of-order ingestion and because each series can potentially be created with samples from up to an hour ago. I chose to detect them based on their first sample.

If the first sample of a series is from the last minute, then we consider the series fresh and don't compare it. This is more robust than comparing with maxT of the query because maxT is often in the future.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>