Fix assertion about matching no values
the `if m.Matches("")` for regular expressions was wrong because it was being run in a function whose return was assumed to be agnostic of whether the matcher is positive or negative (`estimateRegexMatcherCardinality`). However, calling `m.Matches("")` makes it dependant on the type of matcher (`!~` or `=~`). This means that the negation which we had before (`if !foundMatches {return stats.TotalSeries()} else {return stats.TotalSeries() - matchedSeries}`) was wrong because the number it was negating could have already been negated.

This is obviously too complicated, so i also deleted the `foundMatches` logic in this commit.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>