MQE: introduce type for query plan version (#13119)
#### What this PR does

This PR is a follow up to
https://github.com/grafana/mimir/pull/13017#discussion_r2435574207.

It introduces a dedicated type for query plan versions.

I have chosen not to add a changelog entry given this is a refactoring
with no user-visible impact.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.