Don't close pooled query-frontend client if requests fail (#13123)
#### What this PR does

This PR fixes an issue with the querier to query-frontend interaction,
specifically how we pool connections.

We pool the connections used to send results back from queriers to
query-frontends. On the happy path, all responses destined for the same
query-frontend use the same client instance.

If any method on that client instance returns an error, we currently
remove that client instance from the pool by calling `RemoveClient`,
which then immediately closes the client instance. Any other requests
using that same client instance fail. Subsequent requests for the same
frontend trigger the creation of a new client instance. The pool also
periodically checks the health of each client instance and closes it if
it is not healthy.

However, there are a couple of problems with this:

- The method may have returned an error but this doesn't necessarily
indicate an issue with the client instance and the client may be
perfectly usable for other requests. For example, the query-frontend may
have returned an error specific to that request, the payload being sent
to the frontend may exceed the maximum allowed gRPC message size etc.
- Because we retry sending unary responses to query-frontends, if the
response is somehow problematic (eg. it exceeds the maximum allowed gRPC
message size, or the frontend consistently returns an error because the
query has been cancelled), then this can cause other requests to the
same frontend to fail over and over again because the client for the
frontend is closed every time the problematic request is retried.
- For streaming responses, once we've sent the first part of the
response, we can't retry, so the problems above get worse: another
request may close the stream, which means the entire request must be
retried.

I don't believe closing the client is necessary if there is something
wrong with the underlying connection: [the docs for
`ClientConn`](https://pkg.go.dev/google.golang.org/grpc#ClientConn) say
that it will automatically handle errors on established connections and
reestablish the connection, if it can.

Everywhere else in Mimir that uses connection pooling does not call
`RemoveClient`, and relies on health checks to remove irrecoverably
broken client instances, so I propose to do the same here.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/5591,
https://github.com/grafana/mimir/pull/6713

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.