Upgrade dskit (#13142)
#### What this PR does

This PR upgrades dskit to bring in
https://github.com/grafana/dskit/pull/761,
https://github.com/grafana/dskit/pull/760 and
https://github.com/grafana/dskit/pull/759.

I haven't added a changelog entry for
https://github.com/grafana/dskit/pull/761 given it is a docs-only
change, and I haven't added a changelog entry for
https://github.com/grafana/dskit/pull/760 given it is a minor addition
of some new metrics.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.