Move continuous-test `-target` CLI flag to `continuous_test_args` in Jsonnet (#13144)
#### What this PR does

This is a follow-up to #13097. In that PR, the `-target=continuous-test`
CLI flag was added directly to the `continuous_test_container`, rather
than being added to `continuous_test_args`.

This PR moves the CLI flag to `continuous_test_args`, so it can be
overridden by consumers of the library, and the `-target=continous-test`
flag isn't duplicated if consumers have already switched to using the
main Mimir image and have set the flag in `continuous_test_args`
themselves.

#### Which issue(s) this PR fixes or relates to

#13097

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.