fix: TestMimir - include local wlan0 interface
The existing interfaces cover typical Linux wired and MacOS wireless
(possibly wired) devices.

The Linux loopback interface `lo` doesn't appear to be used.

On a Linux laptop without a wired interface `eth0` but
with the wireless interface `wlan0` the following test fails on `main`
but passes with this commit:
```sh
$ go test ./... -run=TestMimir/-target=all,alertmanager

<snip>

--- FAIL: TestMimir (0.01s)
    --- FAIL: TestMimir/-target=all,alertmanager (0.01s)
        mimir_test.go:218:
            	Error Trace:	/<snip>/karlskewes/mimir/pkg/mimir/mimir_test.go:218
            	Error:      	Received unexpected error:
            	           	no useable address found for interfaces [en0 eth0 lo0 lo]
            	           	invalid ring lifecycler config
            	           	github.com/grafana/mimir/pkg/storegateway.newStoreGateway
            	           		/<snip>/karlskewes/mimir/pkg/storegateway/gateway.go:158
```