Clarify error message returned by query-frontends when queriers send a response for an unknown query ID (#13145)
#### What this PR does

This PR adds another possible explanation for why a query with the
provided query ID can't be found.

I've chosen not to add a changelog entry given this is a minor change to
an error message.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.