ingester: mirror querier: ignore fresh series (#13137)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does


Series created between running the first query with index planning and
running the second query without index planning may show up in the
result and create a discrepancy. I call these fresh series.

Detecting fresh series isn't trivial because of out-of-order ingestion
and because each series can potentially be created with samples from up
to an hour ago. I chose to detect them based on their first sample.

If the first sample of a series is from the last minute, then we
consider the series fresh and don't compare it. This is more robust than
comparing with maxT of the query because maxT is often in the future.


#### Which issue(s) this PR fixes or relates to

related to #12460 https://github.com/grafana/mimir/issues/12085

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>