Add zone-aware routing support to memberlist (#13129)
#### What this PR does

In this PR I'm adding experimental support for memberlist zone-aware
routing, in order to reduce cross-AZ data transfer.
[Here](https://github.com/grafana/dskit/blob/main/kv/memberlist/DESIGN.md#zone-aware-routing)
you can read how it works.

The implementation has been done in `memberlist` and `dskit` vendored
libraries. In this PR I'm just updating the vendored libs, plugging the
config, and exposing memberlist node role and zone to the UI.

As part of this work, I also noticed that when we run Mimir with
`-target=memberlist-kv` it doesn't actually run memberlist under the
hood, because memberlist client gets lazily initialized at the first
usage. I've took the opportunity to fix it in this PR too (it's the
change done in `pkg/mimir/modules.go`).

#### Which issue(s) this PR fixes or relates to

Part of https://github.com/grafana/mimir-squad/issues/2634

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>