continuous-test: Do not fail record correctness test when fully empty metadata are dropped (#13158)
#### What this PR does

This fixes a false-positive when a request contains a completely empty
metadata entry. That is, it may be bound to a metric, but the
type/unit/help are all default.

Such metadata are elided at serialization time:
https://github.com/grafana/mimir/blob/2033393ef903341b7fd08c683e922b0ad1c951f3/pkg/mimirpb/mimir.pb.go#L11828

The tester should not fail a record when these are dropped - we are
dropping it intentionally, as these objects have no meaning, so no need
to waste space sending them.

#### Which issue(s) this PR fixes or relates to

contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.