index lookup planning: add a label cardinality threshold for using smaller count-min sketches (#13160)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does
Adds a flag
`blocks-storage.tsdb.index-lookup-planning.label-cardinality-for-smaller-sketch`.
Labels with below this number of series will be allocated a
significantly smaller count-min sketch (i.e. allow a high relative
error), since we don't mind being more off for small values.

#### Which issue(s) this PR fixes or relates to

Relates to #11920 

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.