MQE: include plan version in analysis endpoint response (#13168)
#### What this PR does

This PR expands the output from the experimental query analysis endpoint
to include the query plan version.

I haven't added a changelog entry given this endpoint is very
experimental, not documented, and not intended for end-user use at this
time.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.