Replace default otel labels with xref to Cloud docs (#13115)
This pull request updates the documentation for configuring the
OpenTelemetry Collector in Mimir by removing the section about working
with default OpenTelemetry labels and replacing it with a reference to
the Grafana Cloud documentation. This streamlines the content and
directs users to the most up-to-date information.

#### Which issue(s) this PR fixes or relates to

Fixes https://github.com/grafana/mimir-squad/issues/3270

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.