Jsonnet: refactoring to expose function to create ruler, query-frontend and ruler-query-frontend ScaledObject (#13176)
#### What this PR does

In this PR I'm proposing a refactoring to expose function to create
ruler, query-frontend and ruler-query-frontend ScaledObject. The
generated jsonnet is the exact same, but it allows downstream projects
to easily create custom scaled objects (e.g. zonal ones, which is what
I'm working on, and will be later upstreamed too).

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>